requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pydantic[email]>=2.10.6
typing-extensions>=4.9.0
SQLAlchemy>=2.0.37
psycopg2>=2.9.10dea
requests>=2.32.3
python-jose>=3.4.0
boto3>=1.37.5
python-multipart>=0.0.2s
httpx>=0.28.1